refactor(core): Move frontend settings to service (no-changelog)
